backend/requirements.txt
# Real-Time Recommender System - Backend Requirements
# FastAPI application dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# ML and Data Processing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
lightgbm==4.1.0

# Recommendation Algorithms
implicit==0.7.2  # Matrix Factorization (ALS)

# Embeddings and Vector Search
sentence-transformers==2.3.1
qdrant-client==1.7.0
faiss-cpu==1.7.4

# Feature Store
redis==5.0.1

# MLflow
mlflow==2.9.2
boto3==1.34.14

# Monitoring and Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0
python-json-logger==2.0.7
psutil==5.9.8

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development Tools
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
